Method for delete all working with autorization header

+++ Services/DebtService.cs
@@ -2,6 +2,9 @@
 using System.Net.Http.Json;
 using System.Text.Json;
 using System.Text;
+using Microsoft.JSInterop;
+using System.Net.Http.Headers;
+
 
 
 namespace PairXpensesFS.Services
@@ -9,10 +12,12 @@ namespace PairXpensesFS.Services
     public class DebtService
     {
         private readonly HttpClient _httpClient;
+        private readonly IJSRuntime _jsRuntime;
 
-        public DebtService(HttpClient httpClient)
+        public DebtService(HttpClient httpClient, IJSRuntime jsRuntime)
         {
             _httpClient = httpClient;
+            _jsRuntime = jsRuntime;
         }
 
         public async Task<List<DebtReq>> GetDebtsByUserAsync(int id)
@@ -134,6 +139,18 @@ namespace PairXpensesFS.Services
 
         public async Task<string> DeleteAllDebtsAsync()
         {
+        // Get token from local storage
+            var token = await _jsRuntime.InvokeAsync<string>("localStorage.getItem", "token");
+
+            // Check if token is present
+            if (string.IsNullOrWhiteSpace(token))
+            {
+                throw new InvalidOperationException("Token not found in local storage.");
+            }
+
+            // Add token to authorization header
+            _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+
             HttpResponseMessage response = await _httpClient.DeleteAsync("api/Debt/deleteall");
 
             if (response.IsSuccessStatusCode)
